+++ src/app/pharos-main/analyze-list/sequence-search/sequence-search.component.spec.ts
@@ -1,0 +1,52 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { SequenceSearchComponent } from './sequence-search.component';
+import {ActivatedRoute} from '@angular/router';
+import {MOCKACTIVATEDROUTE} from '../../../../../test/mock-activate-route';
+import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
+import {SharedModule} from '../../../shared/shared.module';
+import {GenericTableModule} from '../../../tools/generic-table/generic-table.module';
+import {RadarChartModule} from '../../../tools/visualizations/radar-chart/radar-chart.module';
+import {RouterTestingModule} from '@angular/router/testing';
+import {AngularFireModule} from '@angular/fire/compat';
+import {COMMON_CONFIG} from '../../../../../test/test-config';
+import {PharosApiService} from '../../../pharos-services/pharos-api.service';
+import {AngularFireAuth} from '@angular/fire/compat/auth';
+import {AngularFirestore} from '@angular/fire/compat/firestore';
+import {FIRESTORESTUB} from '../../../../../test/firestore-stub';
+
+describe('SequenceAlignmentsComponent', () => {
+  let component: SequenceSearchComponent;
+  let fixture: ComponentFixture<SequenceSearchComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ SequenceSearchComponent ],
+      imports: [
+        BrowserAnimationsModule,
+        SharedModule,
+        GenericTableModule,
+        RadarChartModule,
+        RouterTestingModule,
+        AngularFireModule.initializeApp(COMMON_CONFIG)
+      ],
+      providers: [
+        PharosApiService,
+        AngularFireAuth,
+        {provide: ActivatedRoute, useValue: MOCKACTIVATEDROUTE},
+        { provide: AngularFirestore, useValue: FIRESTORESTUB }
+      ],
+    })
+    .compileComponents();
+  });
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(SequenceSearchComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});