+++ src/app/pharos-main/data-details/disease-details/gwas-disease-analytics/gwas-disease-analytics.component.ts
@@ -1,6 +1,5 @@
-import {Component, Input, OnInit} from '@angular/core';
+import {Component, Inject, Input, OnInit, PLATFORM_ID} from '@angular/core';
 import {DynamicPanelComponent} from '../../../../tools/dynamic-panel/dynamic-panel.component';
-import {NavSectionsService} from '../../../../tools/sidenav-panel/services/nav-sections.service';
 import {PharosProperty} from '../../../../models/pharos-property';
 import {ScatterPlotData} from '../../../../tools/visualizations/scatter-plot/scatter-plot.component';
 import {takeUntil} from 'rxjs/operators';
@@ -8,6 +7,7 @@ import {ScatterOptions} from '../../../../tools/visualizations/scatter-plot/mode
 import {PharosPoint} from '../../../../models/pharos-point';
 import {Disease} from '../../../../models/disease';
 import {DynamicServicesService} from '../../../../pharos-services/dynamic-services.service';
+import {isPlatformBrowser} from '@angular/common';
 
 @Component({
   selector: 'pharos-gwas-disease-analytics',
@@ -17,6 +17,7 @@ import {DynamicServicesService} from '../../../../pharos-services/dynamic-servic
 export class GwasDiseaseAnalyticsComponent extends DynamicPanelComponent implements OnInit {
 
   constructor(
+    @Inject(PLATFORM_ID) private platformID: any,
     public dynamicServices: DynamicServicesService) {
     super(dynamicServices);
   }
@@ -100,64 +101,65 @@ export class GwasDiseaseAnalyticsComponent extends DynamicPanelComponent impleme
 
   scatterPlotData: ScatterPlotData[] = [];
 
-  ngOnInit(): void {this._data
-    // listen to data as long as term is undefined or null
-    // Unsubscribe once term has value
-    .pipe(
-      takeUntil(this.ngUnsubscribe)
-    )
-    .subscribe(x => {
-      this.disease = this.data.diseases;
-      this.diseaseProps = this.data.diseasesProps;
+  ngOnInit(): void {
+    this._data
+      // listen to data as long as term is undefined or null
+      // Unsubscribe once term has value
+      .pipe(takeUntil(this.ngUnsubscribe))
+      .subscribe(x => {
+        this.disease = this.data.diseases;
+        this.diseaseProps = this.data.diseasesProps;
 
-      const betaCountPlot = new ScatterPlotData();
-      betaCountPlot.data = [];
-      betaCountPlot.selected = true;
-      betaCountPlot.options = new ScatterOptions({
-        xLabel: 'Mean Rank Score',
-        yLabel: 'Beta Count',
-        yBuffer: 1,
-        margin: {top: 20, right: 35, bottom: 50, left: 50}
-      });
-
-      const orPlot = new ScatterPlotData();
-      orPlot.data = [];
-      orPlot.options = new ScatterOptions({
-        xLabel: 'Mean Rank Score',
-        yLabel: 'Odds Ratio',
-        yBuffer: 1,
-        margin: {top: 20, right: 35, bottom: 50, left: 50}
-      });
-
-      this.scatterPlotData = [betaCountPlot, orPlot];
-      if (this.disease?.gwasAnalytics?.associations.length > 0) {
-        this.showSection();
-        this.disease.gwasAnalytics.associations.forEach(assoc => {
-          if (assoc.medianOddsRatio) {
-            const p: PharosPoint = new PharosPoint({
-              label: assoc.target.gene,
-              x: assoc.meanRankScore,
-              y: assoc.medianOddsRatio,
-              name: assoc.target.name
+        if (isPlatformBrowser(this.platformID)) {
+          if (this.disease?.gwasAnalytics?.associations.length > 0) {
+            this.showSection();
+            const betaCountPlot = new ScatterPlotData();
+            betaCountPlot.data = [];
+            betaCountPlot.selected = true;
+            betaCountPlot.options = new ScatterOptions({
+              xLabel: 'Mean Rank Score',
+              yLabel: 'Beta Count',
+              yBuffer: 1,
+              margin: {top: 20, right: 35, bottom: 50, left: 50}
             });
-            orPlot.data.push(p);
-          }
-          if (assoc.meanRankScore) {
-            const p: PharosPoint = new PharosPoint({
-              label: assoc.target.gene,
-              x: assoc.meanRankScore,
-              y: assoc.betaCount,
-              name: assoc.target.name
-            });
-            betaCountPlot.data.push(p);
-          }
-        });
-      } else {
-        this.hideSection();
-      }
 
-      this.loadingComplete();
-    });
+            const orPlot = new ScatterPlotData();
+            orPlot.data = [];
+            orPlot.options = new ScatterOptions({
+              xLabel: 'Mean Rank Score',
+              yLabel: 'Odds Ratio',
+              yBuffer: 1,
+              margin: {top: 20, right: 35, bottom: 50, left: 50}
+            });
+
+            this.scatterPlotData = [betaCountPlot, orPlot];
+            this.disease.gwasAnalytics.associations.forEach(assoc => {
+              if (assoc.medianOddsRatio) {
+                const p: PharosPoint = new PharosPoint({
+                  label: assoc.target.gene,
+                  x: assoc.meanRankScore,
+                  y: assoc.medianOddsRatio,
+                  name: assoc.target.name
+                });
+                orPlot.data.push(p);
+              }
+              if (assoc.meanRankScore) {
+                const p: PharosPoint = new PharosPoint({
+                  label: assoc.target.gene,
+                  x: assoc.meanRankScore,
+                  y: assoc.betaCount,
+                  name: assoc.target.name
+                });
+                betaCountPlot.data.push(p);
+              }
+            });
+          } else {
+            this.hideSection();
+          }
+        }
+
+        this.loadingComplete();
+      });
   }
 
 }