+++ src/app/tools/knowledge-table/knowledge-table.component.ts
@@ -54,20 +54,10 @@ export class KnowledgeTableComponent extends DynamicPanelComponent implements On
     this._data
     // listen to data as long as term is undefined or null
     // Unsubscribe once term has value
-      .pipe(
-        takeUntil(this.ngUnsubscribe)
-      )
+      .pipe(takeUntil(this.ngUnsubscribe))
       .subscribe(x => {
         this.tableData = this.data.sort((a, b) => b.value.term - a.value.term).slice(0, 5);
         this.changeRef.markForCheck();
       });
-  }
-
-  /**
-   * clean up on leaving component
-   */
-  ngOnDestroy() {
-    this.ngUnsubscribe.next();
-    this.ngUnsubscribe.complete();
   }
 }